+++ .gitignore
@@ -1,2 +1,5 @@
 .idea/
 notebooks/.ipynb_checkpoints/
+# Byte-compiled / optimized / DLL files
+__pycache__/
+
