+++ requirements.txt
@@ -1,0 +1,5 @@
+numpy==1.25.1
+pandas==2.0.3
+scikit-learn==1.0.2
+matplotlib==3.7.2
+joblib==1.3.1
